+++ packages/quantic/coveo.d.ts
@@ -1,5 +1,9 @@
 /* eslint-disable node/no-unpublished-import */
-import {InsightEngine, SearchEngine} from '@coveo/headless';
+import {
+  InsightEngine,
+  SearchEngine,
+  RecommendationEngine,
+} from '@coveo/headless';
 import {LightningElement} from 'lwc';
 import * as BuenoTypes from './force-app/main/default/staticresources/coveobueno/definitions/index';
 import {CoreEngine} from './force-app/main/default/staticresources/coveoheadless/definitions/app/engine';
@@ -7,17 +11,20 @@ import {ExternalEngineOptions} from './force-app/main/default/staticresources/co
 import * as HeadlessCaseAssistTypes from './force-app/main/default/staticresources/coveoheadless/definitions/case-assist.index';
 import * as HeadlessTypes from './force-app/main/default/staticresources/coveoheadless/definitions/index';
 import * as HeadlessInsightTypes from './force-app/main/default/staticresources/coveoheadless/definitions/insight.index';
+import * as HeadlessRecommendationTypes from './force-app/main/default/staticresources/coveoheadless/definitions/recommendation.index';
 
 export * from './force-app/main/default/staticresources/coveoheadless/definitions/index';
 export * from './force-app/main/default/staticresources/coveoheadless/definitions/case-assist.index';
 export * from './force-app/main/default/staticresources/coveoheadless/definitions/insight.index';
+export * from './force-app/main/default/staticresources/coveoheadless/definitions/recommendation.index';
 export * from './force-app/main/default/staticresources/coveobueno/definitions/index';
 
 interface Bindings {
   engine?:
     | HeadlessTypes<CoreEngine>
     | HeadlessCaseAssistTypes<CoreEngine>
-    | HeadlessInsightTypes<CoreEngine>;
+    | HeadlessInsightTypes<CoreEngine>
+    | HeadlessRecommendationTypes<CoreEngine>;
   store?: Record<String, unknown>;
 }
 
@@ -26,12 +33,14 @@ declare global {
   const CoveoHeadless: typeof HeadlessTypes;
   const CoveoHeadlessCaseAssist: typeof HeadlessCaseAssistTypes;
   const CoveoHeadlessInsight: typeof HeadlessInsightTypes;
+  const CoveoHeadlessRecommendation: typeof HeadlessRecommendationTypes;
   type AnyHeadless =
     | CoveoHeadless
     | CoveoHeadlessCaseAssist
-    | CoveoHeadlessInsight;
+    | CoveoHeadlessInsight
+    | CoveoHeadlessRecommendation;
 
-  type AnyEngine = SearchEngine | InsightEngine;
+  type AnyEngine = SearchEngine | InsightEngine | RecommendationEngine;
 
   interface Window {
     coveoHeadless: {